+++ multi-env.d.ts
@@ -1,0 +1,4 @@
+// For import intellisense
+import multiEnv from './dist/multi-env'
+export * from './dist/multi-env'
+export default multiEnv
