+++ src/index.ts
@@ -1,15 +1,10 @@
-import { build as viteBuild, version } from 'vite'
-import type { Plugin, ConfigEnv, UserConfig } from 'vite'
+import { build as viteBuild } from 'vite'
+import type { Plugin } from 'vite'
 
+import { createElectronPlugin } from './base'
 import { triggerStartup } from './startup'
 import type { OnStartOptions } from './startup'
-import {
-  resolveServerUrl,
-  resolveViteConfig,
-  resolveInput,
-  setupMockHtml,
-  withExternalBuiltins,
-} from './utils'
+import { resolveViteConfig, withExternalBuiltins } from './utils'
 
 // public utils
 export { startup } from './startup'
@@ -30,107 +25,48 @@ export function build(options: ElectronOptions): ReturnType<typeof viteBuild> {
 
 export default function electron(options: ElectronOptions | ElectronOptions[]): Plugin[] {
   const optionsArray = Array.isArray(options) ? options : [options]
-  let userConfig: UserConfig
-  let configEnv: ConfigEnv
-  let cleanupMock: (() => Promise<void>) | undefined
+  return createElectronPlugin({
+    prefix: 'vite-plugin-electron',
+    dev(pluginContext, server) {
+      const entryCount = optionsArray.length
+      let closeBundleCount = 0
 
-  if (Number.parseInt(version) < 8) {
-    throw new Error(
-      `[vite-plugin-electron] Vite v${version} does not support \`rolldownOptions\`, please install \`vite@>=8\` or use \`vite-plugin-electron@0.29.1\`.`,
-    )
-  }
+      for (const options of optionsArray) {
+        options.vite ??= {}
+        options.vite.mode ??= server.config.mode
+        options.vite.root ??= server.config.root
+        options.vite.envDir ??= server.config.envDir
+        options.vite.envPrefix ??= server.config.envPrefix
 
-  return [
-    {
-      name: 'vite-plugin-electron:dev',
-      apply: 'serve',
-      configResolved(config) {
-        // When there is no entry (no index.html and no configured input), write a
-        // temporary mock so that Vite's dev server starts without errors.
-        if (!resolveInput(config)) {
-          cleanupMock = setupMockHtml(config, false, config.logger)
+        options.vite.build ??= {}
+        if (!Object.keys(options.vite.build).includes('watch')) {
+          // #252
+          options.vite.build.watch = {}
         }
-      },
-      configureServer(server) {
-        server.httpServer?.once('close', async () => {
-          if (cleanupMock) {
-            await cleanupMock()
-            cleanupMock = undefined
-          }
-        })
+        options.vite.build.minify ??= false
 
-        server.httpServer?.once('listening', () => {
-          Object.assign(process.env, {
-            VITE_DEV_SERVER_URL: resolveServerUrl(server),
-          })
-
-          const entryCount = optionsArray.length
-          let closeBundleCount = 0
-
-          for (const options of optionsArray) {
-            options.vite ??= {}
-            options.vite.mode ??= server.config.mode
-            options.vite.root ??= server.config.root
-            options.vite.envDir ??= server.config.envDir
-            options.vite.envPrefix ??= server.config.envPrefix
-
-            options.vite.build ??= {}
-            if (!Object.keys(options.vite.build).includes('watch')) {
-              // #252
-              options.vite.build.watch = {}
+        options.vite.plugins ??= []
+        options.vite.plugins.push({
+          name: ':startup',
+          closeBundle() {
+            if (++closeBundleCount < entryCount) {
+              return
             }
-            options.vite.build.minify ??= false
-
-            options.vite.plugins ??= []
-            options.vite.plugins.push({
-              name: ':startup',
-              closeBundle() {
-                if (++closeBundleCount < entryCount) {
-                  return
-                }
-                triggerStartup(this, server, options)
-              },
-            })
-            build(options)
-          }
+            triggerStartup(pluginContext, server, options)
+          },
         })
-      },
+        build(options)
+      }
     },
-    {
-      name: 'vite-plugin-electron:prod',
-      apply: 'build',
-      config(config, env) {
-        userConfig = config
-        configEnv = env
-
-        // Make sure that Electron can be loaded into the local file using `loadFile` after packaging.
-        config.base ??= './'
-      },
-      configResolved(config) {
-        // When there is no entry (no index.html and no configured input), write a
-        // temporary mock so that Vite's build has a valid entry point.
-        if (!resolveInput(config)) {
-          cleanupMock = setupMockHtml(config, true, config.logger)
-        }
-      },
-      async closeBundle() {
-        try {
-          for (const options of optionsArray) {
-            options.vite ??= {}
-            options.vite.mode ??= configEnv.mode
-            options.vite.root ??= userConfig.root
-            options.vite.envDir ??= userConfig.envDir
-            options.vite.envPrefix ??= userConfig.envPrefix
-            await build(options)
-          }
-        } finally {
-          // Remove mock files created in configResolved before building Electron.
-          if (cleanupMock) {
-            await cleanupMock()
-            cleanupMock = undefined
-          }
-        }
-      },
+    async build(userConfig, configEnv) {
+      for (const options of optionsArray) {
+        options.vite ??= {}
+        options.vite.mode ??= configEnv.mode
+        options.vite.root ??= userConfig.root
+        options.vite.envDir ??= userConfig.envDir
+        options.vite.envPrefix ??= userConfig.envPrefix
+        await build(options)
+      }
     },
-  ]
+  })
 }
